requirements.txt
requests
pytest
allure-pytest
python-dotenv